TranslateShape Fix
Major fix to the translate shape statistic for 'rectangle' that previously had a typo. Should change the output of all statistics using this setting. The old version was potentially relevant (time-reversal asymmetric at looking at range into the future), but this is how the feature was meant to be constructed.